removed deprecated code

+++ wsa-cachepurge/lib/wsa-cachepurge_wp-module.class.php
@@ -144,7 +144,7 @@ class WSA_Cachepurge_WP
 	 * purging is activated in the plugin.
 	 * 
 	 * @since    1.0.0
-	 * @version  1.1.0
+	 * @version  1.1.2
 	 * @return void
 	 */
 	public static function purge_hooks()
@@ -153,29 +153,9 @@ class WSA_Cachepurge_WP
 		// Fetch the auto purge setting
 		$autoPurge = get_option('wsa-cachepurge_auto-purge');
 
-		// Add switch to update the previous string setting to the new integer format
-		// This will be deprecated and removed in futur version
-		switch ($autoPurge) {
-			
-			case 'on':
-				// Update setting to int '1'
-				update_option('wsa-cachepurge_auto-purge', 1);
-
-				//purge the user cache
-				WSAHandler\WSA::purge_cache();
-				break;
-
-			case 'off':
-
-				// Update setting to int '0'
-				update_option('wsa-cachepurge_auto-purge', 0);
-				break;
-
-			default:
-
-				// Purge the user cache
-				WSAHandler\WSA::purge_cache();
-				break;
+		// Clear the WSA cache
+		if ($autoPurge) {
+			WSAHandler\WSA::purge_cache();
 		}
 	}
 
@@ -240,7 +220,7 @@ class WSA_Cachepurge_WP
 	 * Function to purge the cache with WP hooks included
 	 * 
 	 * @since    1.0.1
-	 * @version  1.1.0
+	 * @version  1.1.2
 	 * @return void
 	 */
 	public static function purge_cache()
@@ -248,24 +228,6 @@ class WSA_Cachepurge_WP
 
 		// Event hook before purging the cache
 		do_action('wsa-cachepurge_before_cache_purge');
-
-		// Fetch the auto purge setting to update the WP options to the new module version.
-		// This will be deprecated and removed in futur version
-		$autoPurge = get_option('wsa-cachepurge_auto-purge');
-
-		// Add switch to update the previous string setting to the new integer format
-		switch ($autoPurge) {
-			
-			case 'on':
-				// Update setting to int '1'
-				update_option('wsa-cachepurge_auto-purge', 1);
-				break;
-
-			case 'off':
-				// Update setting to int '0'
-				update_option('wsa-cachepurge_auto-purge', 0);
-				break;
-		}
 
 		// Call the purge cache function from the WSA class
 		WSAHandler\WSA::purge_cache();
